Add label support introduced in V0.17
Cf.
https://bitcoincore.org/en/releases/0.17.0/#label-and-account-apis-for-wallet,
some methods have been deprecated. To ensure backwards compatibility,
the deprecated methods are not removed in this commit.

The gist of the changes is that the "account" concept is replaced by
the "label" concept which map one-to-one to labels used in the
GUI version of Bitcoin Core. Thus the CLI use of labels is
compatible with GUI use of labels. Previously this was not the case.

See the above link for a description of the new methods.

Added methods:
- GetAddressInfo
- GetAddressByLabel
- GetReceivedByLabel
- ListLabels
- ListReceivedByLabel:
  Like ListReceivedByLabel but the field InvolvesWatchOnly is added. It
  indicates if the returned balance includes tx outputs sent to
  watch-only addresses. This field is null if the reveived amount is 0.
- SetLabel

Changed methods:
- ListUnspent:
  Added a "Label" field, but kept the "Account" field for backwards
  compatibility. Both cases are handled in the associated response
  class' ToString method.

The method arguments named "account" have not been renamed to label.
This may be appropriate to do for v0.18 or if the backwards compatibility
is removed from this library.