Add support for getblock verbosity above 1
In the Bitcoin Core RPC, getblock exists in two version:
one where the 2nd argument is a boolean, and one where it
is an integer where 0 corresponds to false, 1 to true, but
where 2 corresponds to a return value with more verbosity
than true. Specifically, the transaction information, along
with the transaction hex is returned with verbosity = 2.

If someone where to make a block explorer using this library,
it is neccessary to get all of a block's transactions in one
RPC call since this would save a tremendous amount of time.

This solution has been tested locally against my own testnet
node.